+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 ### 2
 * Added Mac support
 * Code cleanup
+
+### 3
+* Added basic windows support
+* mocha tests